discord-bot/requirements.txt
discord
meshtastic
requests
pytz
pubsub
sqlalchemy
psycopg2-binary
pytest
pytest-asyncio
freezegun
mock